refactor: split out check transformation functions to their own check files
